TEX-82: Created BuilderLayout and applied to timeline page
